Version 1.2
Easy File Traversering has one less wrap.  There has also been improvements and changes to the file traversing. Now , the Directories are included so the unit tests where changed to 13.  You need to filter aways those from the results when calculating hashes Stefan (see PR for code).  It should be noted that Java Nio File Traversing is not a good for ransom products. Since we are not making that here  , and we are making a legitime app , there is no real issue for us. Just saying , NIO and Ransomware is not a thing I would Recommend.  However , for Legetime apps , Java NIO seven days in the week <3