Version 1.2.future
Still things that needs improvement but it works good for our use case (hashes for a monitored directory). It works better now than before :) Sticking with one folder and its sub folders might be a good strategy. Best wishes for the weeding concert with your guitar and voice :) Take pictures (or ask the audience) :)